refactor: centralize POSTHOG_HOST validation into env.ts
Convert next.config.mjs to next.config.ts and import the shared env
validation instead of manual checks. Add POSTHOG_HOST to the server
schema in env.ts so all env validation lives in one place.

Closes #285
